refactor away repetition
